+++ include/dctl/evaluate/feature.hpp
@@ -27,8 +27,8 @@ public:
         static int material(Position<Rules, Board> const& p)
         {
                 return
-                        Weight<Rules, Board>::material[0] * p.pieces(Color).count() +
-                        Weight<Rules, Board>::material[1] * p.kings(Color).count()
+                        Weight<Rules, Board>::material[0] * static_cast<int>(p.pieces(Color).count()) +
+                        Weight<Rules, Board>::material[1] * static_cast<int>(p.kings(Color).count())
                 ;
         }
 
@@ -37,7 +37,7 @@ public:
         {
                 int score = 0;
                 for (auto i = 1; i < Board::height; ++i)
-                        score += Weight<Rules, Board>::tempo[i] * (p.pieces(Color) & board::Row<Board>::mask(Color, i)).count();
+                        score += Weight<Rules, Board>::tempo[i] * static_cast<int>((p.pieces(Color) & board::Row<Board>::mask(Color, i)).count());
                 return score;
         }
 
@@ -48,8 +48,8 @@ public:
                 for (auto i = 1; i < Board::width / 2; ++i) {
                         score += Weight<Rules, Board>::center[i] *
                         (
-                                (p.pieces(Color) & board::Column<Board>::mask( Color, i)).count() +
-                                (p.pieces(Color) & board::Column<Board>::mask(!Color, i)).count()
+                                static_cast<int>((p.pieces(Color) & board::Column<Board>::mask( Color, i)).count()) +
+                                static_cast<int>((p.pieces(Color) & board::Column<Board>::mask(!Color, i)).count())
                         );
                 }
                 return score;
@@ -62,8 +62,8 @@ public:
                 for (auto i = 0; i < Board::width / 2; ++i) {
                         score += Weight<Rules, Board>::balance[i] *
                         (
-                                (p.pieces(Color) & board::Column<Board>::mask( Color, i)).count() -
-                                (p.pieces(Color) & board::Column<Board>::mask(!Color, i)).count()
+                                static_cast<int>((p.pieces(Color) & board::Column<Board>::mask( Color, i)).count()) -
+                                static_cast<int>((p.pieces(Color) & board::Column<Board>::mask(!Color, i)).count())
                         );
                 }
                 return -abs(score);
@@ -72,7 +72,7 @@ public:
         template<template<class, class> class Position, class Rules, class Board>
         static int mobility(Position<Rules, Board> const& p)
         {
-                return Weight<Rules, Board>::mobility * successor::mobility<Color>(p);
+                return Weight<Rules, Board>::mobility * static_cast<int>(successor::mobility<Color>(p));
         }
 };
